fix: convert opportunity.amount (common schema) to float (#1366)
- remove trunc after we migrate customer tables

## Test Plan

[Describe test plan here]

## Deployment instructions

[Add any special deployment instructions here]

+++ packages/core/destination_writers/postgres.ts
@@ -733,7 +733,7 @@ CREATE ${temp ? 'TEMP TABLE' : 'TABLE'} IF NOT EXISTS ${
   "last_modified_at" TIMESTAMP(3) NOT NULL,
   "name" TEXT,
   "description" TEXT,
-  "amount" INTEGER,
+  "amount" FLOAT,
   "stage" TEXT,
   "status" TEXT,
   "close_date" TIMESTAMP(3),
